+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN bookdown VERSION 0.23
 
+- `bs4_book()` has now some `<meta>` tags that allows sharing a published book on social media. `cover-image`, `url`, `title` and `description` set in YAML will be used in `index.html` and then modified to be adapted per HTML page (#1034). 
+
 - Style change in `bs4_book()` where code block inside callout blocks will have their background fill the whole width of the bordered block (#1175).
 
 - In `bs4_book()`, math in footnotes is now rendered (@mine-cetinkaya-rundel, #1026)